requirements.txt
aiofiles==24.1.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.114.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.28.1
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pytest==8.3.4
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
